Factory pattern spring DI
