+++ securebanking-openbanking-uk-rcs-api/src/main/java/com/forgerock/securebanking/openbanking/uk/rcs/api/dto/consent/details/DomesticStandingOrderConsentDetails.java
@@ -15,7 +15,6 @@
  */
 package com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details;
 
-import com.forgerock.securebanking.common.openbanking.uk.forgerock.datamodel.account.FRAccountWithBalance;
 import com.forgerock.securebanking.common.openbanking.uk.forgerock.datamodel.common.FRAmount;
 import com.forgerock.securebanking.common.openbanking.uk.forgerock.datamodel.payment.FRWriteDomesticStandingOrderDataInitiation;
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.forgerock.FRFrequency;
@@ -29,10 +28,8 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.joda.time.Instant;
 
-import java.util.List;
-
-import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.intent.members.AMOUNT;
-import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.intent.members.CURRENCY;
+import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.Intent.Members.*;
+import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.Intent.OB_INTENT_OBJECT;
 import static com.forgerock.securebanking.openbanking.uk.rcs.converters.UtilConverter.isNotNull;
 
 /**
@@ -44,29 +41,65 @@ import static com.forgerock.securebanking.openbanking.uk.rcs.converters.UtilConv
 @NoArgsConstructor
 public class DomesticStandingOrderConsentDetails extends ConsentDetails {
     private FRWriteDomesticStandingOrderDataInitiation standingOrder;
-    private List<FRAccountWithBalance> accounts;
     private String paymentReference;
     private FRAmount charges;
+
+    @Override
+    public void mapping(JsonObject consentDetails) {
+
+        if (!consentDetails.has(OB_INTENT_OBJECT)) {
+            throw new IllegalStateException("Expected " + OB_INTENT_OBJECT + " field in json");
+        } else {
+            final JsonObject obIntentObject = consentDetails.get(OB_INTENT_OBJECT).getAsJsonObject();
+            final JsonElement consentDataElement = obIntentObject.get(DATA);
+            if (isNotNull(consentDataElement)) {
+                JsonObject data = consentDataElement.getAsJsonObject();
+
+                if (isNotNull(data.get(INITIATION))) {
+                    JsonObject initiation = data.getAsJsonObject(INITIATION);
+
+                    paymentReference = isNotNull(initiation.get(REFERENCE))
+                            ? initiation.get(REFERENCE).getAsString() : null;
+
+                    setStandingOrder(
+                            isNotNull(initiation.get(FINAL_PAYMENT_DATETIME))
+                                    ? initiation.get(FINAL_PAYMENT_DATETIME) : null,
+                            isNotNull(initiation.get(FINAL_PAYMENT_AMOUNT))
+                                    ? initiation.getAsJsonObject(FINAL_PAYMENT_AMOUNT) : null,
+                            isNotNull(initiation.get(FIRST_PAYMENT_DATETIME))
+                                    ? initiation.get(FIRST_PAYMENT_DATETIME) : null,
+                            isNotNull(initiation.get(FIRST_PAYMENT_AMOUNT))
+                                    ? initiation.getAsJsonObject(FIRST_PAYMENT_AMOUNT) : null,
+                            isNotNull(initiation.get(RECURRING_PAYMENT_DATETIME))
+                                    ? initiation.get(RECURRING_PAYMENT_DATETIME) : null,
+                            isNotNull(initiation.get(RECURRING_PAYMENT_AMOUNT))
+                                    ? initiation.getAsJsonObject(RECURRING_PAYMENT_AMOUNT) : null,
+                            initiation.get(FREQUENCY)
+                    );
+
+                    if (isNotNull(data.get(CHARGES))) {
+                        setCharges(data.getAsJsonArray(CHARGES));
+                    }
+                }
+            }
+        }
+    }
 
     public void setStandingOrder(FRWriteDomesticStandingOrderDataInitiation standingOrder) {
         this.standingOrder = standingOrder;
     }
 
     public void setCharges(JsonArray charges) {
-        if (!isNotNull(charges)) {
-            this.charges = null;
-        } else {
-            this.charges = new FRAmount();
-            Double amount = 0.0;
+        this.charges = new FRAmount();
+        Double amount = 0.0;
 
-            for (JsonElement charge : charges) {
-                JsonObject chargeAmount = charge.getAsJsonObject().getAsJsonObject(AMOUNT);
-                amount += chargeAmount.get(AMOUNT).getAsDouble();
-            }
-
-            this.charges.setCurrency(standingOrder.getFirstPaymentAmount().getCurrency());
-            this.charges.setAmount(amount.toString());
+        for (JsonElement charge : charges) {
+            JsonObject chargeAmount = charge.getAsJsonObject().getAsJsonObject(AMOUNT);
+            amount += chargeAmount.get(AMOUNT).getAsDouble();
         }
+
+        this.charges.setCurrency(standingOrder.getFirstPaymentAmount().getCurrency());
+        this.charges.setAmount(amount.toString());
     }
 
     public void setStandingOrder(
