--- securebanking-openbanking-uk-rcs-api/src/main/java/com/forgerock/securebanking/openbanking/uk/rcs/converters/DomesticStandingOrderConsentDetailsConverter.java
@@ -1,101 +1,0 @@
-/*
- * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.securebanking.openbanking.uk.rcs.converters;
-
-import com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.DomesticScheduledPaymentConsentDetails;
-import com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.DomesticStandingOrderConsentDetails;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.intent.OAUTH2_CLIENT_NAME;
-import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.intent.OB_INTENT_OBJECT;
-import static com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetailsConstants.intent.members.*;
-import static com.forgerock.securebanking.openbanking.uk.rcs.converters.UtilConverter.isNotNull;
-
-/**
- * Converter class to map {@link JsonObject} to {@link DomesticScheduledPaymentConsentDetails}
- */
-@Slf4j
-@NoArgsConstructor
-public class DomesticStandingOrderConsentDetailsConverter {
-
-    private static volatile DomesticStandingOrderConsentDetailsConverter instance;
-
-    /*
-     * Double check locking principle to ensure that only one instance 'DomesticPaymentConsentDetailsConverter' is created
-     */
-    public static DomesticStandingOrderConsentDetailsConverter getInstance() {
-        if (instance == null) {
-            synchronized (DomesticStandingOrderConsentDetailsConverter.class) {
-                if (instance == null) {
-                    instance = new DomesticStandingOrderConsentDetailsConverter();
-                }
-            }
-        }
-        return instance;
-    }
-
-    public DomesticStandingOrderConsentDetails mapping(JsonObject consentDetails) {
-        DomesticStandingOrderConsentDetails details = new DomesticStandingOrderConsentDetails();
-
-        if (!consentDetails.has(OB_INTENT_OBJECT)) {
-            throw new IllegalStateException("Expected " + OB_INTENT_OBJECT + " field in json");
-        } else {
-            final JsonObject obIntentObject = consentDetails.get(OB_INTENT_OBJECT).getAsJsonObject();
-            final JsonElement consentDataElement = obIntentObject.get(DATA);
-            if (!isNotNull(consentDataElement)) {
-                details.setPaymentReference(null);
-                details.setStandingOrder(null);
-            } else {
-                JsonObject data = consentDataElement.getAsJsonObject();
-
-                if (isNotNull(data.get(INITIATION))) {
-                    JsonObject initiation = data.getAsJsonObject(INITIATION);
-
-                    details.setPaymentReference(isNotNull(initiation.get(REFERENCE)) ?
-                            initiation.get(REFERENCE).getAsString() : null);
-
-                    details.setStandingOrder(
-                            isNotNull(initiation.get(FINAL_PAYMENT_DATETIME))
-                                    ? initiation.get(FINAL_PAYMENT_DATETIME) : null,
-                            isNotNull(initiation.get(FINAL_PAYMENT_AMOUNT))
-                                    ? initiation.getAsJsonObject(FINAL_PAYMENT_AMOUNT) : null,
-                            isNotNull(initiation.get(FIRST_PAYMENT_DATETIME))
-                                    ? initiation.get(FIRST_PAYMENT_DATETIME) : null,
-                            isNotNull(initiation.get(FIRST_PAYMENT_AMOUNT))
-                                    ? initiation.getAsJsonObject(FIRST_PAYMENT_AMOUNT) : null,
-                            isNotNull(initiation.get(RECURRING_PAYMENT_DATETIME))
-                                    ? initiation.get(RECURRING_PAYMENT_DATETIME) : null,
-                            isNotNull(initiation.get(RECURRING_PAYMENT_AMOUNT))
-                                    ? initiation.getAsJsonObject(RECURRING_PAYMENT_AMOUNT) : null,
-                            initiation.get(FREQUENCY)
-                    );
-
-                    details.setCharges(isNotNull(data.get(CHARGES)) ?
-                            data.getAsJsonArray(CHARGES) :
-                            null);
-                }
-            }
-        }
-        return details;
-    }
-
-    public final DomesticStandingOrderConsentDetails toDomesticStandingOrderConsentDetails(JsonObject consentDetails) {
-        return mapping(consentDetails);
-    }
-}
