+++ securebanking-openbanking-uk-rcs-api/src/main/java/com/forgerock/securebanking/openbanking/uk/rcs/converters/general/ConsentDetailsFactory.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.forgerock.securebanking.openbanking.uk.rcs.converters.general;
+
+import com.forgerock.securebanking.openbanking.uk.rcs.api.dto.consent.details.ConsentDetails;
+import com.forgerock.securebanking.platform.client.IntentType;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.EnumMap;
+import java.util.List;
+
+@Component
+public class ConsentDetailsFactory {
+
+    private EnumMap<IntentType, ConsentDetails> consentDetailsEnumMap;
+
+    @Autowired
+    public ConsentDetailsFactory(List<ConsentDetails> consentDetailsList) {
+        this.consentDetailsEnumMap = new EnumMap<>(IntentType.class);
+        for (ConsentDetails consentDetails : consentDetailsList) {
+            this.consentDetailsEnumMap.put(consentDetails.getIntentType(), consentDetails);
+        }
+    }
+
+    public ConsentDetails getConsentDetails(IntentType intentType){
+        return this.consentDetailsEnumMap.get(intentType);
+    }
+}
